+++ pearl_qt.py
@@ -1,3 +1,4 @@
+# pearl_qt.py
 import sys
 import logging
 from src.utils.dependencies import check_dependencies