+++ src/interface/settings/advanced_settings.py
@@ -3,6 +3,8 @@ import customtkinter as ctk
 from src.utils.save_settings import save_all_settings
 import src.utils.fonts as fonts
 from CTkToolTip import CTkToolTip
+from tkinter import messagebox
+from PySide6.QtWidgets import QMessageBox
 from src.utils.save_settings import load_data_path
 from src.connections.ollama import uninstall_ollama
 from src.connections.docker import uninstall_docker
@@ -11,6 +13,7 @@ from src.utils.factory_reset import total_factory_reset
 import subprocess
 import os
 import logging
+import sys
 
 
 def create_advanced_tab(globals, advanced_frame):
@@ -225,4 +228,35 @@ def create_advanced_tab(globals, advanced_frame):
 
     ctk.CTkButton(save_button_frame,
                   text="Save Settings",
-                  command=lambda: save_all_settings(globals)).pack()
+                  command=lambda: save_button(globals)).pack()
+
+    def save_button(globals):
+        """Saves and prompts for restart if required."""
+        prompt_restart = False
+        if globals.github_check != globals.github_check_var.get():
+            prompt_restart = True
+        elif globals.active_theme != globals.theme_var.get():
+            prompt_restart = True
+        elif globals.beta != globals.beta_var.get():
+            prompt_restart = True
+        save_all_settings(globals)
+
+        if prompt_restart:
+            if globals.qt_mode:
+                reply = QMessageBox.question(
+                    None,
+                    "Restart Pearl?",
+                    f"Would you like to restart Pearl to apply all changes?",
+                    QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
+                    QMessageBox.StandardButton.Yes)
+                if reply == QMessageBox.StandardButton.Yes:
+                    subprocess.Popen(globals.app_path)
+                    sys.exit(0)
+            else:
+                    reply = messagebox.askyesno(
+                        parent=globals.root,
+                        title="Restart Pearl",
+                        message="Would you like restart Pearl to apply all changes?")
+                    if reply:
+                        subprocess.Popen(globals.app_path)
+                        sys.exit(0)